# infrastructure/inbound/rest/src/main/resources/api.yml
openapi: 3.0.1
info:
  title: 'Inventory OpenAPI spec'
  version: 0.0.6
servers:
  - url: http://localhost:8080
    description: local
paths:
  /caps:
    get:
      tags:
        - cap-controller
      operationId: findCaps
      parameters:
        - name: country
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cap'
    put:
      tags:
        - cap-controller
      operationId: upsertCap
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertCap'
        required: true
      responses:
        '204':
          description: No-content
  /caps/{id}:
    get:
      tags:
        - cap-controller
      operationId: findCapById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Cap'
        '404':
          description: Cap not found
    delete:
      tags:
        - cap-controller
      operationId: deleteCapById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: Cap not found
  /coutries:
    get:
      tags:
        - country-controller
      operationId: findCountries
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Country'
  /countries/{code}:
    get:
      tags:
        - country-controller
      operationId: findCountryByCode
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Country'
        '404':
          description: Country not found
  /books:
    get:
      tags:
        - book-controller
      operationId: findBooks
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
    put:
      tags:
        - book-controller
      operationId: upsertBook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Book'
        required: true
      responses:
        '204':
          description: No-content
  /books/{id}:
    get:
      tags:
        - book-controller
      operationId: findBookById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Book'
        '404':
          description: Book not found
    delete:
      tags:
        - book-controller
      operationId: deleteBookById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
        '404':
          description: Book not found
components:
  schemas:
    UpsertCap:
      type: object
      properties:
        business_id:
          type: string
        name:
          type: string
        description:
          type: string
        amount:
          type: integer
          format: int32
        country:
          type: string
    Cap:
      type: object
      properties:
        business_id:
          type: string
        name:
          type: string
        description:
          type: string
        amount:
          type: integer
          format: int32
        country:
          $ref: '#/components/schemas/Country'
    Country:
      type: object
      properties:
        name:
          type: string
        code:
          type: string
    Book:
      type: object
      properties:
        business_id:
          type: string
        title:
          type: string
        author:
          type: string
        description:
          type: string